requirements.txt
numpy>=2.0
pandas>=2.0
PyYAML>=6.0
pyEDFlib>=0.1.42
torch>=1.12
